requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
brisque==0.0.15
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
distro==1.9.0
exceptiongroup==1.2.0
fastapi==0.109.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
imageio==2.33.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.0
langchain-community==0.0.12
langchain-core==0.1.10
langchain-openai==0.0.2.post1
langsmith==0.0.80
lazy_loader==0.3
libsvm==3.23.0.4
marshmallow==3.20.2
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numpy==1.26.3
openai==1.7.2
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
pydantic==2.5.3
pydantic_core==2.14.6
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-image==0.22.0
scipy==1.11.4
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
starlette==0.35.1
tenacity==8.2.3
tifffile==2023.12.9
tiktoken==0.5.2
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
urllib3==1.26.18
uvicorn==0.25.0
yarl==1.9.4
